requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-filter==24.2
django-tenants==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.7
h11==0.14.0
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
simplejwt==2.0.1
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.30.1
